+++ classes/models/NostoOrder.php
@@ -24,6 +24,7 @@
  */
 
 use Nosto\Object\Order\Order as NostoSDKOrder;
+use Nosto\Object\Cart\LineItem as NostoSDKLineItem;
 
 class NostoOrder extends NostoSDKOrder
 {
@@ -114,24 +115,17 @@ class NostoOrder extends NostoSDKOrder
         $totalShippingTaxIncl = 0;
         $totalWrappingTaxIncl = 0;
         $totalGiftTaxIncl = 0;
+        $totalProductTaxIncl = 0;
 
         // One order can be split into multiple orders, so we need to combine their data.
         $orderCollection = Order::getByReference($order->reference);
         foreach ($orderCollection as $item) {
             /** @var $item Order */
             $products = array_merge($products, $item->getProducts());
-            $totalDiscountsTaxIncl = NostoHelperPrice::roundPrice(
-                $totalDiscountsTaxIncl + $item->total_discounts_tax_incl,
-                $currency
-            );
-            $totalShippingTaxIncl = NostoHelperPrice::roundPrice(
-                $totalShippingTaxIncl + $item->total_shipping_tax_incl,
-                $currency
-            );
-            $totalWrappingTaxIncl = NostoHelperPrice::roundPrice(
-                $totalWrappingTaxIncl + $item->total_wrapping_tax_incl,
-                $currency
-            );
+            $totalDiscountsTaxIncl += $item->total_discounts_tax_incl;
+            $totalShippingTaxIncl += $item->total_shipping_tax_incl;
+            $totalWrappingTaxIncl += $item->total_wrapping_tax_incl;
+            $totalProductTaxIncl += $item->total_products_wt;
         }
 
         // We need the cart rules used for the order to check for gift products and free shipping.
@@ -161,10 +155,7 @@ class NostoOrder extends NostoSDKOrder
                         } else {
                             $productPriceWt = 0;
                         }
-                        $totalGiftTaxIncl = NostoHelperPrice::roundPrice(
-                            $totalGiftTaxIncl + $productPriceWt,
-                            $currency
-                        );
+                        $totalGiftTaxIncl += $productPriceWt;
                         $giftProduct = $product;
                         $giftProduct['product_quantity'] = 1;
                         $giftProduct['product_price_wt'] = 0;
@@ -198,6 +189,7 @@ class NostoOrder extends NostoSDKOrder
 
                 $purchasedItem = new NostoOrderPurchasedItem();
                 $purchasedItem->setProductId((string)$p->id);
+                $purchasedItem->setSkuId((string)$idAttribute);
                 $purchasedItem->setQuantity((int)$item['product_quantity']);
                 $purchasedItem->setName((string)$productName);
                 $purchasedItem->setPrice($item['product_price_wt']);
@@ -211,17 +203,14 @@ class NostoOrder extends NostoSDKOrder
 
             if ($totalDiscountsTaxIncl > 0) {
                 // Subtract possible gift product price from total as gifts are tagged with price zero (0).
-                $totalDiscountsTaxIncl = NostoHelperPrice::roundPrice(
-                    $totalDiscountsTaxIncl - $totalGiftTaxIncl,
-                    $currency
-                );
+                $totalDiscountsTaxIncl -= $totalGiftTaxIncl;
                 if ($totalDiscountsTaxIncl > 0) {
                     $purchasedItem = new NostoOrderPurchasedItem();
-                    $purchasedItem->setProductId("-1");
+                    $purchasedItem->setProductId(NostoSDKLineItem::PSEUDO_PRODUCT_ID);
                     $purchasedItem->setQuantity(1);
                     $purchasedItem->setName('Discount');
                     // Note the negative value.
-                    $purchasedItem->setPrice(-$totalDiscountsTaxIncl);
+                    $purchasedItem->setPrice(NostoHelperPrice::roundPrice(-$totalDiscountsTaxIncl, $currency));
                     $purchasedItem->setPriceCurrencyCode((string)$currency->iso_code);
                     $purchasedItems[] = $purchasedItem;
                 }
@@ -238,22 +227,40 @@ class NostoOrder extends NostoSDKOrder
                 }
             }
 
+            if ($freeShipping && $totalShippingTaxIncl > 0 && $totalDiscountsTaxIncl > 0) {
+                $totalDiscountsTaxIncl -= $totalShippingTaxIncl;
+            }
+
+            //deduct the gift product among from product among
+            $totalProductTaxIncl -= $totalGiftTaxIncl;
+            $discountPercentage = max(0, $totalDiscountsTaxIncl / $totalProductTaxIncl);
+
+            /** @var NostoOrderPurchasedItem $purchasedItem */
+            foreach ($purchasedItems as $purchasedItem) {
+                if ($purchasedItem->getProductId() === NostoSDKLineItem::PSEUDO_PRODUCT_ID) {
+                    continue;
+                }
+
+                $unitPrice = $purchasedItem->getUnitPrice() * (1 - $discountPercentage);
+                $purchasedItem->setPrice(NostoHelperPrice::roundPrice($unitPrice, $currency));
+            }
+
             if (!$freeShipping && $totalShippingTaxIncl > 0) {
                 $purchasedItem = new NostoOrderPurchasedItem();
-                $purchasedItem->setProductId("-1");
+                $purchasedItem->setProductId(NostoSDKLineItem::PSEUDO_PRODUCT_ID);
                 $purchasedItem->setQuantity(1);
                 $purchasedItem->setName('Shipping');
-                $purchasedItem->setPrice($totalShippingTaxIncl);
+                $purchasedItem->setPrice(NostoHelperPrice::roundPrice($totalShippingTaxIncl, $currency));
                 $purchasedItem->setPriceCurrencyCode((string)$currency->iso_code);
                 $purchasedItems[] = $purchasedItem;
             }
 
             if ($totalWrappingTaxIncl > 0) {
                 $purchasedItem = new NostoOrderPurchasedItem();
-                $purchasedItem->setProductId("-1");
+                $purchasedItem->setProductId(NostoSDKLineItem::PSEUDO_PRODUCT_ID);
                 $purchasedItem->setQuantity(1);
                 $purchasedItem->setName('Gift Wrapping');
-                $purchasedItem->setPrice($totalWrappingTaxIncl);
+                $purchasedItem->setPrice(NostoHelperPrice::roundPrice($totalWrappingTaxIncl, $currency));
                 $purchasedItem->setPriceCurrencyCode((string)$currency->iso_code);
                 $purchasedItems[] = $purchasedItem;
             }