+++ controllers/front/reloadCart.php
@@ -23,12 +23,10 @@
  * @license   http://opensource.org/licenses/afl-3.0.php  Academic Free License (AFL 3.0)
  */
 
-require_once(dirname(__FILE__) . '/api.php');
-
 /**
  * Front controller for reload cart tagging
  */
-class NostoTaggingReloadCartModuleFrontController extends NostoTaggingApiModuleFrontController
+class NostoTaggingReloadCartModuleFrontController extends ModuleFrontController
 {
     /**
      * @inheritdoc