+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to Semantic Versioning(http://semver.org/).
 
+## 3.1.0
+- Support restoring abandoned shopping cart
+- Add product features to product custom field tagging
+- Add sku id to cart and order tagging
+
 ## 3.0.3
 - Change product tag1, tag2 and tag3 tagging to tags1, tags2 and tags3 to compatible with nosto backend